Move HIGH_LATENCY2 stream to separate file and update some comments"
